Move parser_obj.py to _internal
